pyproject.toml
[project]
name = "english-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "blinker==1.9.0",
    "bottle==0.13.4",
    "cachetools==5.5.2",
    "certifi==2025.8.3",
    "cffi==1.17.1",
    "charset-normalizer==3.4.3",
    "click==8.2.1",
    "clr-loader==0.2.7.post0",
    "colorama==0.4.6",
    "flask==3.1.1",
    "google-ai-generativelanguage==0.6.15",
    "google-api-core==2.25.1",
    "google-api-python-client==2.179.0",
    "google-auth==2.40.3",
    "google-auth-httplib2==0.2.0",
    "google-cloud-speech==2.33.0",
    "google-cloud-texttospeech==2.27.0",
    "google-generativeai==0.8.5",
    "googleapis-common-protos==1.70.0",
    "grpcio==1.74.0",
    "grpcio-status==1.71.2",
    "httplib2==0.22.0",
    "idna==3.10",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "markupsafe==3.0.2",
    "proto-plus==1.26.1",
    "protobuf==5.29.5",
    "proxy-tools==0.1.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.22",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pyparsing==3.2.3",
    "python-dotenv==1.1.1",
    "pythonnet==3.0.5",
    "pywebview==6.0",
    "pywin32==311",
    "requests==2.32.4",
    "rsa==4.9.1",
    "tqdm==4.67.1",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "uritemplate==4.2.0",
    "urllib3==2.5.0",
    "werkzeug==3.1.3",
]
